+++ translator/src/main/scala/scopus/translator/io/package.scala
@@ -1,0 +1,7 @@
+package scopus.translator
+
+/**
+ * This packages contains all the I/O classes.
+ */
+package object io {
+}