+++ translator/src/main/scala/scopus/translator/language/package.scala
@@ -1,0 +1,7 @@
+package scopus.translator
+
+/**
+ * This package defines the most representative characteristics of Scopus.
+ */
+package object language {
+}