+++ translator/src/main/scala/scopus/translator/replacement/package.scala
@@ -1,0 +1,7 @@
+package scopus.translator
+
+/**
+ * This package contains auxiliary classes for string manipulation.
+ */
+package object replacement {
+}